requirements.txt
django
djangorestframework
django-fitler
psycopg2-binary
pytest-django